# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GHW API",
    "description": "A lightweight API for MLH Global Hack Week.",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Root",
        "description": "Simple health endpoint.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthCheck"
                }
              }
            }
          }
        }
      }
    },
    "/challenges": {
      "get": {
        "tags": [
          "challenges"
        ],
        "summary": "List Challenges",
        "description": "List all available challenges.",
        "operationId": "list_challenges_challenges_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Challenge"
                  },
                  "type": "array",
                  "title": "Response List Challenges Challenges Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "challenges"
        ],
        "summary": "Create Challenge",
        "description": "Create a new challenge.",
        "operationId": "create_challenge_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/challenges/{challenge_id}": {
      "get": {
        "tags": [
          "challenges"
        ],
        "summary": "Get Challenge",
        "description": "Retrieve a single challenge by ID.",
        "operationId": "get_challenge_challenges__challenge_id__get",
        "parameters": [
          {
            "name": "challenge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Challenge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          },
          "404": {
            "description": "Challenge not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Challenge": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Build a TODO API"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ],
            "default": "easy"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "published": {
            "type": "boolean",
            "title": "Published",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description"
        ],
        "title": "Challenge"
      },
      "ChallengeCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Build a TODO API"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ],
            "default": "easy"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "published": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "ChallengeCreate"
      },
      "Difficulty": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "Difficulty"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthCheck": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok",
            "example": "ok"
          },
          "uptime_seconds": {
            "type": "number",
            "title": "Uptime Seconds"
          }
        },
        "type": "object",
        "required": [
          "uptime_seconds"
        ],
        "title": "HealthCheck"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Health checks and runtime info."
    },
    {
      "name": "challenges",
      "description": "Endpoints to manage hack week challenges."
    }
  ]
}